backend/requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.1.0
pydantic[email]==2.5.2
slowapi==0.1.9
